+++ CHANGELOG.md
@@ -13,6 +13,8 @@ For details about compatibility between different releases, see the **Commitment
 
 ### Changed
 
+- Do not print error line logs for rate limited gRPC and HTTP API requests.
+
 ### Deprecated
 
 ### Removed